fix(class tracking) EVAL-531 fix getting teachers appreciations annually. (#246)

+++ src/main/java/fr/openent/competences/controllers/ExportPDFController.java
@@ -626,7 +626,7 @@ public class ExportPDFController extends ControllerHelper {
 
         Promise<JsonArray> multiTeachingPromise = Promise.promise();
         utilsService.getMultiTeachers(idEtablissement,
-                new JsonArray().add(idClasse), idPeriode.intValue(), FutureHelper.handlerJsonArray(multiTeachingPromise.future()));
+                new JsonArray().add(idClasse), idPeriode, FutureHelper.handlerJsonArray(multiTeachingPromise.future()));
 
         Promise<JsonArray> servicesPromise = Promise.promise();
         utilsService.getServices(idEtablissement,
